requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0
sentence-transformers==2.3.1
faiss-cpu==1.7.4
transformers==4.36.0
torch==2.1.2
wikipedia==1.4.0
numpy==1.26.0
tqdm==4.66.1
python-multipart==0.0.6
